+++ api/client/src/main/java/org/projectnessie/client/auth/oauth2/TokenExchangeFlow.java
@@ -35,9 +35,6 @@ class TokenExchangeFlow extends AbstractFlow {
 
   @Override
   public Tokens fetchNewTokens(Tokens currentTokens) {
-    if (!config.getTokenExchangeEnabled()) {
-      throw new MustFetchNewTokensException("Token exchange is disabled");
-    }
     Objects.requireNonNull(currentTokens);
     TokensExchangeRequest.Builder request =
         TokensExchangeRequest.builder()